Reapply "Merge branch 'dev' into ui_marketplacework"
This reverts commit 1316f555ea72edb451bc0e1cd05edafc1d321c62.
