+++ app/src/main/java/com/agora/app/backend/lambda/KeyNotFoundException.java
@@ -1,4 +1,4 @@
-package com.agora.app.lambda;
+package com.agora.app.backend.lambda;
 
 public class KeyNotFoundException extends RuntimeException {
     public KeyNotFoundException (String message) {
