+++ app/src/main/java/com/agora/app/backend/lambda/Operations.java
@@ -1,4 +1,4 @@
-package com.agora.app.lambda;
+package com.agora.app.backend.lambda;
 
 public enum Operations {
     PUT(true, true),
