--- app/src/main/java/com/agora/app/dynamodb/ChatWrapper.java
@@ -1,4 +1,0 @@
-package com.agora.app.dynamodb;
-
-public class ChatWrapper {
-}
