--- app/src/main/java/com/agora/app/dynamodb/MessageWrapper.java
@@ -1,4 +1,0 @@
-package com.agora.app.dynamodb;
-
-public class MessageWrapper {
-}
