+++ app/src/main/java/com/agora/app/frontend/ExpandedListingActivity.java
@@ -1,10 +1,12 @@
 package com.agora.app.frontend;
+import com.agora.app.backend.base.Listing;
 
 import android.os.Bundle;
 import android.widget.ImageView;
 import android.widget.TextView;
 import androidx.appcompat.app.AppCompatActivity;
 import com.agora.app.R;
+import java.util.ArrayList;
 
 public class ExpandedListingActivity extends AppCompatActivity {
 
@@ -27,33 +29,31 @@ public class ExpandedListingActivity extends AppCompatActivity {
         // Initialize views
         titleTextView = findViewById(R.id.expanded_listing_title);
         descriptionTextView = findViewById(R.id.expanded_listing_description);
-        priceTextView = findViewById(R.id.expanded_listing_price);
+        priceTextView = findViewById(R.id.expanded_listing_price); 
         listingImageView = findViewById(R.id.expanded_listing_image);
         tag1TextView = findViewById(R.id.expanded_listing_tag1);
         tag2TextView = findViewById(R.id.expanded_listing_tag2);
         tag3TextView = findViewById(R.id.expanded_listing_tag3);
 
-
-
         // Get the data passed from the previous activity
         String title = getIntent().getStringExtra("title");
         String description = getIntent().getStringExtra("description");
-        float price = getIntent().getFloatExtra("price", 0.0f);
+        float price = getIntent().getFloatExtra("price", 0.0f); 
         int imageResource = getIntent().getIntExtra("image", R.drawable.ic_placeholder);
-        String tag1 = getIntent().getStringExtra("tag1");
-        String tag2 = getIntent().getStringExtra("tag2");
-        String tag3 = getIntent().getStringExtra("tag3");
+        ArrayList<String> tags = getIntent().getStringArrayListExtra("tags"); 
 
         // Set the values to the respective views
         titleTextView.setText(title);
         descriptionTextView.setText(description);
-        priceTextView.setText("$" + String.format("%.2f", price));
+        priceTextView.setText("$" + String.format("%.2f", price)); // Set price
         listingImageView.setImageResource(imageResource);
-        tag1TextView.setText(tag1);
-        tag2TextView.setText(tag2);
-        tag3TextView.setText(tag3);
 
-
-
+        // Dynamically set the tags (check if tags are not null)
+        if (tags != null) {
+            // Make sure to handle cases where the tag list may be smaller than 3
+            tag1TextView.setText(tags.size() > 0 ? tags.get(0) : "No Tag");
+            tag2TextView.setText(tags.size() > 1 ? tags.get(1) : "No Tag");
+            tag3TextView.setText(tags.size() > 2 ? tags.get(2) : "No Tag");
+        }
     }
 }
