+++ app/src/main/java/com/agora/app/frontend/UserInfoActivity.java
@@ -1,4 +1,5 @@
 package com.agora.app.frontend;
+import com.agora.app.backend.base.Listing;
 
 import com.agora.app.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
@@ -26,6 +27,7 @@ import android.widget.EditText;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.UUID;
 
 /**
  * @class UserInfoActivity
@@ -150,9 +152,17 @@ public class UserInfoActivity extends AppCompatActivity {
             String title = titleInput.getText().toString();
             String description = descriptionInput.getText().toString();
             
+            
             String tag1 = tag1Input.getText().toString();
             String tag2 = tag2Input.getText().toString();
             String tag3 = tag3Input.getText().toString();
+
+            ArrayList<String> tags = new ArrayList<>();
+            if (!tag1.isEmpty()) tags.add(tag1);
+            if (!tag2.isEmpty()) tags.add(tag2);
+            if (!tag3.isEmpty()) tags.add(tag3);
+
+            
 
             float price = 0.0f;
             String priceString = priceInput.getText().toString();
@@ -167,7 +177,13 @@ public class UserInfoActivity extends AppCompatActivity {
             }    
 
             if(!title.isEmpty() && !description.isEmpty()) {
-                Listing newListing = new Listing(title, description, price, tag1, tag2, tag3);
+                UUID uuid = UUID.randomUUID();
+                String displayName = "temp"; 
+                String username = "user"; 
+                String type = "default"; 
+
+                Listing newListing = new Listing(uuid, title, price, description, displayName, username, type, tags);
+    
                 ListingManager.getInstance().addListing(newListing);
                 view.notifyItemInserted(listings.size() - 1);
                 dialog.dismiss();
