+++ app/src/main/res/drawable/blue_border.xml
@@ -1,0 +1,4 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <stroke android:width="5dp" android:color="#003071"/>
+    <solid android:color="#FFFFFF"/>
+</shape>
