+++ app/src/main/res/layout/account_creator.xml
@@ -2,7 +2,21 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:padding="64dp">
+
+    <!--page title-->
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:padding="16dp"
+        android:text="Create an Account"
+        android:textSize="25sp"
+        android:textColor="#003071"
+        android:gravity="center"
+        android:layout_marginBottom="50dp"
+        android:background="@drawable/blue_border"/>
 
     <EditText
         android:id="@+id/username_id"
@@ -26,15 +40,13 @@
         android:layout_width="match_parent"
         android:autofillHints="password"
         android:inputType="text"
-        android:hint="Please Confirm Your Password..." />
+        android:hint="Confirm Password" />
 
     <Button
         android:id="@+id/button_id"
         android:layout_width="match_parent"
         android:layout_height="56dp"
         android:layout_marginTop="32dp"
-        android:text="Submit"
-        android:textColor="@color/white"
-        android:backgroundTint="@color/light_green"/>
+        android:text="Submit"/>
 
 </LinearLayout>
