+++ app/src/main/res/layout/activity_marketplace.xml
@@ -3,22 +3,47 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <!--page title-->
-    <TextView
+    <!-- Spacer to move page below phone's top information bar -->
+    <View
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Marketplace"
-        android:textSize="40sp"
-        android:gravity="center"/>
+        android:layout_height="50dp"
+        android:background="#003071"/>
 
-    <!--page buffer-->
+    <!-- Header -->
+    <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="?attr/actionBarSize"
+            android:gravity="center_horizontal">
+
+            <ImageView
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:src="@drawable/agora_header"
+                android:layout_gravity="center"
+                android:contentDescription="Agora Header"
+                android:background="#003071"/>
+    </LinearLayout>
+
+    <!-- Page Title -->
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight=".075"
-        android:orientation="vertical">
+        android:layout_height="wrap_content"
+        android:gravity="center_horizontal"
+        android:padding="16dp">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="10dp"
+            android:text="Marketplace"
+            android:textSize="40sp"
+            android:textColor="#003071"
+            android:gravity="center"
+            android:background="@drawable/blue_border"
+            />
     </LinearLayout>
 
     <!--search bar-->
@@ -31,7 +56,10 @@
         app:queryHint="Search"
         android:focusable="false"
         app:searchIcon="@drawable/icon_search"
-        android:background="@drawable/search_bar"/>
+        android:background="@drawable/search_bar"
+        android:layout_marginBottom="16dp"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="16dp"/>
 
     <Button
         android:layout_width="match_parent"
@@ -94,7 +122,7 @@
         android:layout_height="wrap_content"
         android:layout_centerInParent="true"
         android:background="@android:color/holo_blue_dark"
-        android:text="No product listings are currently available :("
+        android:text="No listings are currently available :("
         android:textSize="30sp"
         android:gravity="center"
         />
@@ -105,24 +133,18 @@
         android:layout_height="0dp"
         android:id="@+id/item_listings"
         android:scrollbars="vertical"
-        android:layout_weight="1"/>
-
-    <!--page buffer-->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight=".1"
-        android:orientation="vertical">
-    </LinearLayout>
+        android:layout_weight="1"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="16dp"/>
 
     <!--nav bar-->
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/nav_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="?android:attr/windowBackground"
-        app:itemIconTint="@color/black"
-        app:itemTextColor="@color/black"
+        android:background="#003071"
+        app:itemIconTint="@color/white"
+        app:itemTextColor="@color/white"
         app:menu="@menu/nav_bar" />
 
 </LinearLayout>
