+++ app/src/main/res/layout/activity_messaging.xml
@@ -5,13 +5,45 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <!--page title-->
-    <TextView
+    <!-- Spacer to move page below phone's top information bar -->
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:background="#003071"/>
+
+    <!-- Header -->
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:gravity="center_horizontal">
+
+        <ImageView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:src="@drawable/agora_header"
+            android:layout_gravity="center"
+            android:contentDescription="Agora Header"
+            android:background="#003071"/>
+    </LinearLayout>
+
+    <!-- Page Title -->
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Messaging Page"
-        android:textSize="40sp"
-        android:gravity="center"/>
+        android:gravity="center_horizontal"
+        android:padding="16dp">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="10dp"
+            android:text="Messaging Page"
+            android:textSize="40sp"
+            android:textColor="#003071"
+            android:gravity="center"
+            android:background="@drawable/blue_border"
+            />
+    </LinearLayout>
 
     <TextView
         android:layout_width="match_parent"
@@ -32,9 +64,9 @@
         android:id="@+id/nav_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="?android:attr/windowBackground"
-        app:itemIconTint="@color/black"
-        app:itemTextColor="@color/black"
+        android:background="#003071"
+        app:itemIconTint="@color/white"
+        app:itemTextColor="@color/white"
         app:menu="@menu/nav_bar" />
 
 </LinearLayout>
