+++ app/src/main/res/layout/activity_saved_posts.xml
@@ -4,19 +4,20 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Saved Posts"
-        android:textSize="40sp"
-        android:layout_marginBottom="20dp"/>
-
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/buttonUserInfo"
         android:text="User Info"
-        android:layout_marginTop="20dp"/>
+        android:layout_marginTop="-10dp"/>
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:gravity="center"
+        android:text="Saved Posts"
+        android:textSize="40sp" />
+
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
