+++ app/src/main/res/layout/item_listing.xml
@@ -5,7 +5,7 @@
     android:layout_height="wrap_content"
     android:orientation="vertical"
     android:padding="16dp"
-    android:background="@drawable/box_border"
+    android:background="@drawable/blue_border_rounded"
     android:layout_marginBottom="10dp">
     <ImageView
         android:id="@+id/listing_image"
@@ -30,7 +30,6 @@
             android:text="Listing Title"
             android:layout_gravity="start"
             android:layout_weight="1" />
-
   
         <TextView
             android:id="@+id/listing_price"
