+++ app/src/main/res/layout/login_activity.xml
@@ -2,15 +2,21 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:padding="64dp">
 
     <!--page title-->
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:gravity="center"
+        android:padding="16dp"
         android:text="Log In"
-        android:textSize="40sp"/>
+        android:textSize="25sp"
+        android:textColor="#003071"
+        android:gravity="center"
+        android:layout_marginBottom="50dp"
+        android:background="@drawable/blue_border"/>
 
     <EditText
         android:id="@+id/username_id"
@@ -33,8 +39,6 @@
         android:layout_width="match_parent"
         android:layout_height="56dp"
         android:layout_marginTop="32dp"
-        android:text="Submit"
-        android:textColor="@color/white"
-        android:backgroundTint="@color/teal_700"/>
+        android:text="Submit" />
 
 </LinearLayout>
