+++ app/src/test/java/com/agora/app/TestClass.java
@@ -7,10 +7,10 @@ import com.agora.app.backend.base.ImageChunk;
 import com.agora.app.backend.base.Password;
 import com.agora.app.backend.base.PaymentMethods;
 import com.agora.app.backend.base.User;
-import com.agora.app.dynamodb.DynamoTables;
-import com.agora.app.lambda.KeyNotFoundException;
-import com.agora.app.lambda.LambdaHandler;
-import com.agora.app.lambda.Operations;
+import com.agora.app.backend.lambda.DynamoTables;
+import com.agora.app.backend.lambda.KeyNotFoundException;
+import com.agora.app.backend.lambda.LambdaHandler;
+import com.agora.app.backend.lambda.Operations;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.BeforeClass;
